% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environmental_data_example_2}
\alias{environmental_data_example_2}
\title{Environmental Sample Data - Example 2}
\format{A data frame with 5 environmental variables for 19 samples}
\source{
\url{http://hdl.handle.net/10915/58915}
}
\usage{
data(environmental_data_example_2)
}
\description{
This sample data is a partial dataset from Sathicq, María Belén. (2017). Empleo de descriptores fitoplanctónicos como biomonitores en la evaluación de la calidad del agua en la costa del río de la Plata (Franja Costera Sur). PhD thesis. http://hdl.handle.net/10915/58915
}
\references{
Sathicq, María Belén. (2017). Empleo de descriptores fitoplanctónicos como biomonitores en la evaluación de la calidad del agua en la costa del río de la Plata (Franja Costera Sur). PhD thesis.
}
\keyword{datasets}
