% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environmental_df}
\alias{environmental_df}
\title{Environmental Sample Data}
\format{A data frame with 11 environmental variables for 50 samples}
\source{
\url{http://hdl.handle.net/10915/58915}
}
\usage{
data(environmental_data)
}
\description{
This sample data is a partial dataset from Sathicq, María Belén. (2017). Empleo de descriptores fitoplanctónicos como biomonitores en la evaluación de la calidad del agua en la costa del río de la Plata (Franja Costera Sur). PhD thesis. http://hdl.handle.net/10915/58915
}
\references{
Sathicq, María Belén. (2017). Empleo de descriptores fitoplanctónicos como biomonitores en la evaluación de la calidad del agua en la costa del río de la Plata (Franja Costera Sur). PhD thesis.
}
\keyword{datasets}
