% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_calculate.R
\name{op_calculate}
\alias{op_calculate}
\title{This function of Optimos Prime calculates optima and tolerance for a matrix of species and environmental factors}
\usage{
op_calculate(enviromental_df, species_df, islog10 = FALSE)
}
\arguments{
\item{enviromental_df}{The dataframe with your environmental data. Variables as rows, Sites as columns}

\item{species_df}{The dataframe with your species densities. Species as rows, Sites as columns.}

\item{islog10}{Boolean. If set to 'TRUE' it means that your environmental data is already transformed to log10. Default = False}
}
\description{
You will need two dataframes. If they are not specified as arguments, you will be prompted to import them from CSV format.
The resulting dataframe from the op_calculate() function will be a dataframe of species (rows) and the optima and tolerance range (+ and -) of the environmental variables (columns)
\itemize{
\item Matrix 1: Species (rows) by Sampling sites (columns).
    First row needs to be the sampling sites names.
    First column needs to be the species' names.
    Values in cells therefore need to be the density of each species at each site.
\item Matrix 2: Environmental variables (rows) by Sampling sites (columns).
    First row needs to be the sampling sites names.
    First column needs to be the names of the environmental variables (i.e. physical-chemical parameters).
    Values in cells therefore need to be the value of each environmental variable at each site.
}
The calculations for optima and tolerance ranges is conducted according to the article by Potapova & Charles (2003):
\itemize{
\item Potapova, M., & Charles, D. F. (2003). Distribution of benthic diatoms in US rivers in relation to conductivity and ionic composition. Freshwater Biology, 48(8), 1311-1328.
}
}
\keyword{density}
\keyword{ecology,}
\keyword{optimum,}
\keyword{species}
\keyword{tolerance,}
