% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_plot.R
\name{op_plot}
\alias{op_plot}
\title{Optimos Prime calculates optima and tolerance for a matrix of species and environmental factors}
\usage{
op_plot(optimaDF, label)
}
\arguments{
\item{optimaDF}{The dataframe resulting from the op_calculate() method.}

\item{label}{(optional) The label for the X axis (usually the environmental variable's name)}
}
\description{
This function plots the specified variable in a caterpillar plot
}
\details{
This function (op_plot) generates caterpillar plots for a specified environmental variable showing optima and tolerance ranges as calculated with the op_calculate() function
}
\concept{density}
\concept{ecology}
\concept{optimum}
\concept{species}
\concept{tolerance}
