% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimos.prime.R
\docType{package}
\name{optimos.prime}
\alias{optimos.prime}
\alias{optimos.prime-package}
\title{Optimos Prime: A package for calculating optima and tolerance range for biological species}
\description{
Optimos Prime: A package for calculating optima and tolerance range for biological species
}
\section{Functions}{

op_calculate()
op_lists()
op_plot()
}

