% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_df}
\alias{species_df}
\title{Species Sample Data}
\format{A data frame with the relative abundance for 57 species in 50 samples}
\source{
\url{http://hdl.handle.net/10915/58915}
}
\usage{
data(species_data)
}
\description{
This sample data is a partial dataset from Sathicq, María Belén. (2017). Empleo de descriptores fitoplanctónicos como biomonitores en la evaluación de la calidad del agua en la costa del río de la Plata (Franja Costera Sur). PhD thesis. http://hdl.handle.net/10915/58915
}
\references{
Sathicq, María Belén. (2017). Empleo de descriptores fitoplanctónicos como biomonitores en la evaluación de la calidad del agua en la costa del río de la Plata (Franja Costera Sur). PhD thesis. http://hdl.handle.net/10915/58915
}
\keyword{datasets}
