% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimos.prime.R
\docType{package}
\name{optimos.prime}
\alias{optimos.prime}
\alias{optimos.prime-package}
\title{Optimos Prime: A package for calculating autoecological data (optima and tolerance range) for biological species}
\description{
The calculations for optima and tolerance ranges is conducted according to the article by Birks et al. (1990). A case study for this technique is shown in Potapova & Charles (2003):
\itemize{
\item Birks, H.J.B., Line J.M., Juggins S., Stevenson A.C., and Ter Braak C.J.F. (1990). Diatoms and pH reconstruction. Philosophical Transactions of the Royal Society B 327:263–278
\item Potapova, M., & Charles, D. F. (2003). Distribution of benthic diatoms in US rivers in relation to conductivity and ionic composition. Freshwater Biology, 48(8), 1311-1328.
}
Sample data is taken from:
\itemize{
\item Sathicq, Maria Belen. (2017). Empleo de descriptores fitoplanctonicos como biomonitores en la evaluacion de la calidad del agua en la costa del Rio de la Plata (Franja Costera Sur). PhD thesis. <http://hdl.handle.net/10915/58915>
}
}
\section{Functions}{

op_calculate()
op_lists()
op_plot()
}

